Improve test code coverage
